[maximal_planar] move StackItem to own file
