+++ src/embeding/maximal_planar/mod.rs
@@ -2,3 +2,4 @@ pub mod index;
 mod phase1;
 mod phase2;
 mod phase3;
+mod stack_item;
